search: add predicate package
