+++ internal/search/predicate/types.go
@@ -1,0 +1,23 @@
+package predicate
+
+import (
+	"github.com/sourcegraph/sourcegraph/internal/search/query"
+)
+
+type P interface {
+	// Field is the name of the field that the predicate applies to.
+	// For example, with `file:contains()`, Field returns "file".
+	Field() string
+
+	// Name is the name of the predicate.
+	// For example, with `file:contains()`, Name returns "contains".
+	Name() string
+
+	// ParseParams parses the contents of the predicate arguments
+	// into the predicate object.
+	ParseParams(string) error
+
+	// Plan generates a plan of (possibly multiple) queries to execute the
+	// behavior of a predicate in a parent query.
+	Plan(parent query.Basic) (query.Plan, error)
+}
