+++ internal/search/query/predicates.go
@@ -1,0 +1,8 @@
+package query
+
+// All recognized predicates
+var AllPredicates = map[string]map[string]struct{}{
+	FieldRepo: {
+		"contains": empty,
+	},
+}
